+++ controller/lib/hal/hal_stm32.cpp
@@ -288,7 +288,7 @@ static void Timer6ISR() {
 
 void HalApi::delay(Duration d) {
   int64_t start = msCount;
-  while (static_cast<uint64_t>(msCount - start) < d.milliseconds()) {
+  while (msCount - start < d.milliseconds()) {
   }
 }
